changes to mongodb

+++ domain/kurs/interfaces.go
@@ -11,5 +11,5 @@ type IKursRepository interface {
 	InsertERates(symbol string, eRate ERate, date string) (err error)
 	InsertTTCounter(symbol string, tt TTCounter, date string) (err error)
 	InsertBankNotes(symbol string, note BankNote, date string) (err error)
-	IsExistKurs(data DataKurs) (countBN, countTT, countERates int, err error)
+	IsExistKurs(data DataKurs) (countBN, countTT, countERates int64, err error)
 }
